+++ FredBoat/src/main/java/fredboat/audio/player/GuildPlayer.kt
@@ -34,7 +34,7 @@ import fredboat.audio.lavalink.SentinelLink
 import fredboat.audio.queue.*
 import fredboat.audio.queue.limiter.QueueLimitStatus
 import fredboat.audio.queue.limiter.QueueLimiter
-import fredboat.audio.queue.tbd.*
+import fredboat.audio.queue.handlers.*
 import fredboat.command.music.control.VoteSkipCommand
 import fredboat.commandmeta.MessagingException
 import fredboat.commandmeta.abs.CommandContext
