+++ FredBoat/src/main/java/fredboat/audio/queue/handlers/IQueueHandler.kt
@@ -1,4 +1,4 @@
-package fredboat.audio.queue.tbd
+package fredboat.audio.queue.handlers
 
 import fredboat.audio.queue.AudioTrackContext
 import org.bson.types.ObjectId
