+++ FredBoat/src/main/java/fredboat/audio/queue/handlers/IShufflableQueueHandler.kt
@@ -1,4 +1,4 @@
-package fredboat.audio.queue.tbd
+package fredboat.audio.queue.handlers
 
 import fredboat.audio.queue.AudioTrackContext
 import java.util.concurrent.ConcurrentLinkedDeque
